+++ Lmyc/Controllers/API/AccountAPIController.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
+using AspNet.Security.OpenIdConnect.Primitives;
 using Lmyc.Models;
 using Lmyc.Models.AccountViewModels;
 using Lmyc.Services;